// frontend/src/lib/grpc/gen/folder/v1/encryption_pb.ts
// @generated by protoc-gen-es v2.13.0 with parameter "target=ts"
// @generated from file folder/v1/encryption.proto (package folder.v1, syntax proto2)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file folder/v1/encryption.proto.
 */
export const file_folder_v1_encryption: GenFile = /*@__PURE__*/
  fileDesc("Chpmb2xkZXIvdjEvZW5jcnlwdGlvbi5wcm90bxIJZm9sZGVyLnYxIhgKB1ZlcnNpb24SDQoFdmFsdWUYASACKA0iGgoJQWxnb3JpdGhtEg0KBXZhbHVlGAIgAigJInIKDkVuY3J5cHRlZFZhdWx0EgoKAml2GAEgAigJEgsKA3RhZxgCIAIoCRIjCgd2ZXJzaW9uGAMgAigLMhIuZm9sZGVyLnYxLlZlcnNpb24SIgoEYWxnbxgEIAIoCzIULmZvbGRlci52MS5BbGdvcml0aG0iRwoORW5jcnlwdGVkQmxvYnMSJwoEbWV0YRgBIAIoCzIZLmZvbGRlci52MS5FbmNyeXB0ZWRWYXVsdBIMCgRkYXRhGAIgAigM");

/**
 * @generated from message folder.v1.Version
 */
export type Version = Message<"folder.v1.Version"> & {
  /**
   * @generated from field: required uint32 value = 1;
   */
  value: number;
};

/**
 * Describes the message folder.v1.Version.
 * Use `create(VersionSchema)` to create a new message.
 */
export const VersionSchema: GenMessage<Version> = /*@__PURE__*/
  messageDesc(file_folder_v1_encryption, 0);

/**
 * @generated from message folder.v1.Algorithm
 */
export type Algorithm = Message<"folder.v1.Algorithm"> & {
  /**
   * @generated from field: required string value = 2;
   */
  value: string;
};

/**
 * Describes the message folder.v1.Algorithm.
 * Use `create(AlgorithmSchema)` to create a new message.
 */
export const AlgorithmSchema: GenMessage<Algorithm> = /*@__PURE__*/
  messageDesc(file_folder_v1_encryption, 1);

/**
 * @generated from message folder.v1.EncryptedVault
 */
export type EncryptedVault = Message<"folder.v1.EncryptedVault"> & {
  /**
   * Base64 encoded 12 length IV
   * Output string length is 16
   *
   * @generated from field: required string iv = 1;
   */
  iv: string;

  /**
   * Base64 encoded 16 length tag
   * Output string length is 24
   *
   * @generated from field: required string tag = 2;
   */
  tag: string;

  /**
   * @generated from field: required folder.v1.Version version = 3;
   */
  version?: Version | undefined;

  /**
   * @generated from field: required folder.v1.Algorithm algo = 4;
   */
  algo?: Algorithm | undefined;
};

/**
 * Describes the message folder.v1.EncryptedVault.
 * Use `create(EncryptedVaultSchema)` to create a new message.
 */
export const EncryptedVaultSchema: GenMessage<EncryptedVault> = /*@__PURE__*/
  messageDesc(file_folder_v1_encryption, 2);

/**
 * @generated from message folder.v1.EncryptedBlobs
 */
export type EncryptedBlobs = Message<"folder.v1.EncryptedBlobs"> & {
  /**
   * @generated from field: required folder.v1.EncryptedVault meta = 1;
   */
  meta?: EncryptedVault | undefined;

  /**
   * @generated from field: required bytes data = 2;
   */
  data: Uint8Array;
};

/**
 * Describes the message folder.v1.EncryptedBlobs.
 * Use `create(EncryptedBlobsSchema)` to create a new message.
 */
export const EncryptedBlobsSchema: GenMessage<EncryptedBlobs> = /*@__PURE__*/
  messageDesc(file_folder_v1_encryption, 3);